api/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy>=2.0
pydantic>=2.7
pydantic-settings>=2.2
httpx>=0.27
python-multipart>=0.0.9
langgraph>=0.3
langchain-openai>=0.1
litellm>=1.40
icalendar>=5.0
tavily-python>=0.3
nest-asyncio>=1.6
aiosqlite>=0.20
alembic>=1.13
python-dotenv>=1.0
